allow templates

+++ install.pl
@@ -4,6 +4,10 @@ use strict;
 use warnings;
 
 use FindBin;
+use Getopt::Long;
+
+use lib "$FindBin::Bin/lib";
+use Template::Tiny;
 
 
 my %APP = (
@@ -11,6 +15,18 @@ my %APP = (
     git  => [qw/gitconfig/],
     vim  => [qw/vim vimrc/],
 );
+
+
+my $context = 'home';
+my %var = (
+    email => 'uwe@uwevoelker.de',
+);
+my $result  = GetOptions(
+    'context=s' => \$context,
+    'email=s'   => \$var{email},
+);
+# set context
+$var{context} = {$context => 1};
 
 
 unless (@ARGV) {
@@ -21,6 +37,9 @@ unless (@ARGV) {
     exit 1;
 }
 
+
+my $tt = Template::Tiny->new(TRIM => 1);
+
 foreach my $name (@ARGV) {
     unless (exists $APP{$name}) {
         print "Unknown app '$name'\n";
@@ -28,8 +47,26 @@ foreach my $name (@ARGV) {
     }
     my $files = $APP{$name};
     foreach my $file (@$files) {
-        system "rm -rf ~/.$file\n";
-        system "ln -s $FindBin::Bin/$file ~/.$file\n";
+        my $path = "$FindBin::Bin/$file";
+        if (-f "$path.tt") {
+            my $template;
+            {
+                local $/;
+                open(my $fh, '<', "$path.tt") or die "$file.tt: $!";
+                $template = <$fh>;
+                close $fh or die "$file.tt: $!";
+            }
+            my $out;
+            $tt->process(\$template, \%var, \$out);
+            my $out_path = "$ENV{HOME}/.$file";
+            open(my $fh, '>', $out_path) or die "$out_path: $!";
+            print $fh $out               or die "$out_path: $!";
+            close $fh                    or die "$out_path: $!";
+        }
+        else {
+            system "rm -rf ~/.$file\n";
+            system "ln -s $FindBin::Bin/$file ~/.$file\n";
+        }
     }
 }
